src/postcard-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
sqlalchemy
alembic
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx

# 安全组件依赖
bleach
pycryptodome

# 监控组件依赖
psutil
httpx
